requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
beanie==1.25.0
motor==3.4.0
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
aiosmtplib==3.0.1
aiohttp==3.9.5
